Implement dynamic compiler
