+++ crates/cubecl-wgpu/Cargo.toml
@@ -28,6 +28,7 @@ cubecl-core = { path = "../cubecl-core", version = "0.5.0", default-features = f
 cubecl-runtime = { path = "../cubecl-runtime", version = "0.5.0", default-features = false, features = [
     "channel-mutex",
 ] }
+derive_more = { workspace = true }
 sanitize-filename = { workspace = true, optional = true }
 
 # SPIR-V
