+++ crates/cubecl-wgpu/src/backend/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod vulkan;
+mod wgsl;
+
+pub use base::*;
