+++ crates/cubecl-wgpu/src/compiler/mod.rs
@@ -1,5 +1,3 @@
 pub mod base;
-pub mod wgsl;
-
-#[cfg(feature = "spirv")]
 pub mod spirv;
+pub mod wgsl;
